requirements.txt
Babel==2.14.0
blinker==1.7.0
click==8.1.7
colorama==0.4.6
Flask==3.0.3
Flask-Admin==1.6.1
Flask-BabelEx==0.9.4
Flask-SQLAlchemy==3.1.1
greenlet==3.0.3
itsdangerous==2.1.2
Jinja2==3.1.3
MarkupSafe==2.1.5
PyMySQL==1.1.0
speaklater==1.3
SQLAlchemy==2.0.29
typing_extensions==4.11.0
Werkzeug==3.0.2
WTForms==3.1.2

cloudinary~=1.39.1